+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -49,7 +49,7 @@ import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.simona.util.TickUtil.TickLong
 import edu.ie3.util.TimeUtil
 import edu.ie3.util.quantities.QuantityUtils.*
-import edu.ie3.util.scala.quantities.DefaultQuantities.zeroKWh
+import edu.ie3.util.scala.quantities.DefaultQuantities.*
 import edu.ie3.util.scala.quantities.{Kilovars, ReactivePower}
 import org.apache.pekko.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
 import org.apache.pekko.actor.typed.ActorRef
@@ -152,8 +152,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         // first request
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -167,8 +167,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         // second request with same voltage
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -182,7 +182,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
           Each(0.98),
-          Each(0),
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -224,8 +224,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -306,8 +306,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -365,8 +365,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -503,8 +503,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -622,8 +622,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -748,8 +748,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -834,8 +834,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -954,8 +954,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -1018,8 +1018,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -1134,8 +1134,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -1246,8 +1246,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -1452,8 +1452,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -1640,8 +1640,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -1830,8 +1830,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           12 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )
 
@@ -2008,8 +2008,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         participantAgent ! RequestAssetPowerMessage(
           24 * 3600,
-          Each(1),
-          Each(0),
+          onePU,
+          zeroPU,
           gridAgent.ref,
         )